backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.0
requests
Pillow>=9.0.0
django-storages
boto3
django-filter
django-extensions